requirements.txt
flask=2.1.2
google.generativeai=0.3.2
